+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 All notable changes to the `C-mantic` extension will be documented in this file.
 
 ## [Unreleased]
+### Added
+- Added support for explicit template specialization. Definitions for specialized function templates/member functions of specialized class templates can now be generated in the matching source file.
+
 ### Fixed
 - Fixed title of `Add Declaration` code-action displaying the wrong target file.
 - Fixed recognition of explicit template specializations, and multi-lined template statements.